[Refactor] fix relative path
